+++ src/contractual_conditions/dto/create_contractual_conditions.dto.ts
