+++ src/toilet_maintenance/toilet_maintenance.service.ts
@@ -1,11 +1,11 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
-import { UpdateToiletMaintenanceDto } from './dto/update-toilet_maintenance.dto';
+import { UpdateToiletMaintenanceDto } from './dto/update_toilet_maintenance.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ToiletMaintenance } from './entities/toilet_maintenance.entity';
 import { ChemicalToilet } from '../chemical_toilets/entities/chemical_toilet.entity';
 import { Repository } from 'typeorm';
-import { CreateToiletMaintenanceDto } from './dto/create-toilet_maintenance.dto';
-import { FilterToiletMaintenanceDto } from './dto/filter-toilet_maintenance.dto';
+import { CreateToiletMaintenanceDto } from './dto/create_toilet_maintenance.dto';
+import { FilterToiletMaintenanceDto } from './dto/filter_toilet_maintenance.dto';
 
 @Injectable()
 export class ToiletMaintenanceService {
