+++ src/vehicle_maintenance/vehicle_maintenance.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { VehicleMaintenanceRecord } from './entities/vehicle-maintenance-record.entity';
-import { VehicleMaintenanceService } from './vehicle-maintenance.service';
-import { VehicleMaintenanceController } from './vehicle-maintenance.controller';
+import { VehicleMaintenanceRecord } from './entities/vehicle_maintenance_record.entity';
+import { VehicleMaintenanceService } from './vehicle_maintenance.service';
+import { VehicleMaintenanceController } from './vehicle_maintenance.controller';
 import { VehiclesModule } from '../vehicles/vehicles.module';
 import { RolesModule } from '../roles/roles.module';
