+++ src/vehicles/vehicles.service.ts
@@ -7,8 +7,8 @@ import {
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Vehicle } from './entities/vehicle.entity';
-import { CreateVehicleDto } from './dto/create-vehicle.dto';
-import { UpdateVehicleDto } from './dto/update-vehicle.dto';
+import { CreateVehicleDto } from './dto/create_vehicle.dto';
+import { UpdateVehicleDto } from './dto/update_vehicle.dto';
 
 @Injectable()
 export class VehiclesService {
